backend/requirements-render.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.5
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
pillow>=10.0.0
redis>=5.0.0
sqlalchemy>=2.0.20
psycopg2-binary>=2.9.0
prometheus-client>=0.16.0
sentry-sdk>=1.25.0
structlog>=22.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
